# src/LarFinance.Api/OpenApi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "LarFinance API",
    "version": "1.0.0",
    "description": "API para cadastro de pessoas, transações e consulta de totais residenciais."
  },
  "servers": [{ "url": "http://localhost:5080" }],
  "paths": {
    "/api/people": {
      "get": {
        "summary": "Lista todas as pessoas",
        "responses": { "200": { "description": "Pessoas ordenadas por nome" } }
      },
      "post": {
        "summary": "Cadastra uma pessoa",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreatePerson" } } }
        },
        "responses": {
          "201": { "description": "Pessoa cadastrada" },
          "400": { "description": "Dados inválidos" }
        }
      }
    },
    "/api/people/{id}": {
      "delete": {
        "summary": "Exclui uma pessoa e suas transações",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": {
          "204": { "description": "Pessoa e transações removidas" },
          "404": { "description": "Pessoa não encontrada" }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CreatePerson": {
        "type": "object",
        "required": ["name", "age"],
        "properties": {
          "name": { "type": "string", "minLength": 2, "maxLength": 120, "example": "Ana Souza" },
          "age": { "type": "integer", "minimum": 0, "maximum": 130, "example": 27 }
        }
      },
...
    }
  }
}